requirements.txt
gymnasium[classic-control]
gymnasium[other]
numpy
scipy
matplotlib
control
tqdm
ipykernel
stable-baselines3
tqdm
rich
ipywidgets
pybullet
pin-pink
qpsolvers[quadprog]
robot_descriptions
